Refactor AzureBlobStorage logic and enhance scanning support.
Refactored relative path handling in AzureBlobStorage for improved folder processing. Added scanning-specific logic for file and folder operations, ensuring proper handling during scanning processes. These changes enhance functionality and maintain consistency across storage operations.

+++ app/Atro/Services/Record.php
@@ -181,7 +181,7 @@ class Record extends RecordService
             $chunkSize = $minChunkSize;
         } else {
             if ($total >= ($minChunkSize * $maxConcurrentJobs) && $total <= ($maxChunkSize * $maxConcurrentJobs)) {
-                $chunkSize = ceil($total / $maxConcurrentJobs);
+                $chunkSize = (int)ceil($total / $maxConcurrentJobs);
             } else {
                 $chunkSize = $maxChunkSize;
             }
